iOS: policy cleanups for the View as Text sheet
DocC docs on the public TerminalTextSnapshot memberwise initializer and
SelectableTextView moved to its own file per one-major-type-per-file.
